requirements.txt
ffmpeg-python
pyglet
pyglet-ffmpeg2